requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai>=1.3.0
python-dotenv==0.19.0
geopy==2.3.0
requests==2.31.0
httpx>=0.27.0
langchain-openai>=0.0.2
langchain-community>=0.0.1
huggingface-hub>=0.19.4
